Simplify markLoaded - no longer need to pass it in
