+++ src/utils/handlePrintWindowOnLoad.ts
@@ -19,7 +19,6 @@ export interface HandlePrintWindowOnLoadData {
     /** The original canvas nodes. Used to apply paints not copied when cloning the nodes */
     originalCanvasNodes: never[] | NodeListOf<HTMLCanvasElement>;
 };
-type MarkLoaded = (resource: Element | Font | FontFace, errorMessages?: unknown[]) => void;
 
 const DEFAULT_PAGE_STYLE = `
     @page {
@@ -38,11 +37,10 @@ const DEFAULT_PAGE_STYLE = `
 
 /**
  * Handles loading resources into the print window and copying over various DOM elements that
- * require special handling. Continuously checks
+ * require special handling.
  */
 export function handlePrintWindowOnLoad(
     printWindow: HTMLIFrameElement,
-    markLoaded: MarkLoaded,
     data: HandlePrintWindowOnLoadData,
     options: UseReactToPrintOptions
 ) {
@@ -64,6 +62,41 @@ export function handlePrintWindowOnLoad(
         suppressErrors,
         copyShadowRoots
     } = options;
+
+    const resourcesLoaded: (Element | Font | FontFace)[] = [];
+    const resourcesErrored: (Element | Font | FontFace)[] = [];
+
+    function markLoaded(resource: Element | Font | FontFace, errorMessages?: unknown[]) {
+        if (resourcesLoaded.includes(resource)) {
+            logMessages({
+                level: "debug",
+                messages: ["Tried to mark a resource that has already been handled", resource],
+                suppressErrors,
+            });
+            return;
+        }
+
+        if (!errorMessages) {
+            resourcesLoaded.push(resource);
+        } else {
+            logMessages({
+                messages: [
+                    '"react-to-print" was unable to load a resource but will continue attempting to print the page',
+                    ...errorMessages
+                ],
+                suppressErrors,
+            });
+            resourcesErrored.push(resource);
+        }
+
+        // We may have errors, but attempt to print anyways - maybe they are trivial and the
+        // user will be ok ignoring them
+        const numResourcesManaged = resourcesLoaded.length + resourcesErrored.length;
+
+        if (numResourcesManaged === numResourcesToLoad) {
+            startPrint(printWindow, options);
+        }
+    };
 
     // Some agents, such as IE11 and Enzyme (as of 2 Jun 2020) continuously call the
     // `onload` callback. This ensures that it is only called once.
@@ -322,6 +355,7 @@ export function handlePrintWindowOnLoad(
         }
     }
 
+    // Handle the base case where there are no resources to load
     if (numResourcesToLoad === 0) {
         startPrint(printWindow, options);
     }
